+++ src/ChannelServer/Inventory.h
@@ -18,7 +18,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #ifndef INVENTORY_H
 #define INVENTORY_H
 
-#define GETINVENTORY(itemid) (itemid/1000000)
+#define GETINVENTORY(itemid) static_cast<uint8_t>(itemid/1000000)
 #define ISARROW(x) (x/10000 == 206)
 #define ISSTAR(x) (x/10000 == 207)
 #define ISBULLET(x) (x/10000 == 233) // Pirate's bullets and capsules
@@ -27,6 +27,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #define ISPET(x) (x >= 5000000 && x <= 5000053)
 #define GETWEAPONTYPE(x) (x/10000)
 
+#include "Types.h"
 #include <unordered_map>
 #include <vector>
 
@@ -59,104 +60,104 @@ enum {
 
 struct EquipInfo {
 	char slots;
-	int price;
-	short istr;
-	short idex;
-	short iint;
-	short iluk;
-	short ihp;
-	short imp;
-	short iwatk;
-	short imatk;
-	short iwdef;
-	short imdef;
-	short iacc;
-	short iavo;
-	short ihand;
-	short ijump;
-	short ispeed;
-	short tamingmob;
+	int32_t price;
+	int16_t istr;
+	int16_t idex;
+	int16_t iint;
+	int16_t iluk;
+	int16_t ihp;
+	int16_t imp;
+	int16_t iwatk;
+	int16_t imatk;
+	int16_t iwdef;
+	int16_t imdef;
+	int16_t iacc;
+	int16_t iavo;
+	int16_t ihand;
+	int16_t ijump;
+	int16_t ispeed;
+	int16_t tamingmob;
 	bool cash; 
 	bool quest;
 };
 
 struct SummonBag {
-	int mobid;
-	int chance;
+	int32_t mobid;
+	uint32_t chance;
 };
 
 struct Skillbook{
-	int skillid;
-	int reqlevel;
-	int maxlevel;
+	int32_t skillid;
+	uint8_t reqlevel;
+	uint8_t maxlevel;
 };
 
 struct ConsumeInfo {
-	short hp;
-	short mp;
-	short hpr;
-	short mpr;
-	int moveTo;
-	int time;
-	short watk;
-	short matk;
-	short avo;
-	short acc;
-	short wdef;
-	short mdef;
-	short speed;
-	short jump;
-	short morph;
-	int success;
-	int cursed;
-	short istr;
-	short idex;
-	short iint;
-	short iluk;
-	short ihp;
-	short imp;
-	short iwatk;
-	short imatk;
-	short iwdef;
-	short imdef;
-	short iacc;
-	short iavo;
-	short ihand;
-	short ijump;
-	short ispeed;
+	int16_t hp;
+	int16_t mp;
+	int16_t hpr;
+	int16_t mpr;
+	int32_t moveTo;
+	int32_t time;
+	int16_t watk;
+	int16_t matk;
+	int16_t avo;
+	int16_t acc;
+	int16_t wdef;
+	int16_t mdef;
+	int16_t speed;
+	int16_t jump;
+	int16_t morph;
+	uint32_t success;
+	uint32_t cursed;
+	int16_t istr;
+	int16_t idex;
+	int16_t iint;
+	int16_t iluk;
+	int16_t ihp;
+	int16_t imp;
+	int16_t iwatk;
+	int16_t imatk;
+	int16_t iwdef;
+	int16_t imdef;
+	int16_t iacc;
+	int16_t iavo;
+	int16_t ihand;
+	int16_t ijump;
+	int16_t ispeed;
 	unsigned char types[8];
-	vector<short> vals;
+	vector<int16_t> vals;
 	vector<SummonBag> mobs;
 	vector<Skillbook> skills;
 };
 
 struct ItemInfo {
-	short maxslot;
-	int price;
+	int16_t maxslot;
+	int32_t price;
 	bool quest;
 	bool consume;
 	ConsumeInfo cons;
 };
 
 namespace Inventory {
-	extern unordered_map<int, EquipInfo> equips;
-	extern unordered_map<int, ItemInfo> items;
-	void addEquipInfo(int id, EquipInfo equip);
-	void addItemInfo(int id, ItemInfo item);
+	extern unordered_map<int32_t, EquipInfo> equips;
+	extern unordered_map<int32_t, ItemInfo> items;
+	void addEquipInfo(int32_t id, EquipInfo equip);
+	void addItemInfo(int32_t id, ItemInfo item);
 	void startTimer();
-	extern bool isCash(int itemid);
+	extern bool isCash(int32_t itemid);
 	extern ItemTimer * timer;
 	void useShop(Player *player, ReadPacket *packet);
 	void useStorage(Player *player, ReadPacket *packet);
 	void itemMove(Player *player, ReadPacket *packet);
-	short addItem(Player *player, Item *item, bool is = false);
-	void addNewItem(Player *player, int itemid, int amount);
-	void takeItem(Player *player, int itemid, int howmany);
+	int16_t addItem(Player *player, Item *item, bool is = false);
+	void addNewItem(Player *player, int32_t itemid, int16_t amount);
+	void takeItem(Player *player, int32_t itemid, uint16_t howmany);
 	void useItem(Player *player, ReadPacket *packet);
 	void cancelItem(Player *player, ReadPacket *packet);
-	void endItem(Player *player, int itemid);
+	void endItem(Player *player, int32_t itemid);
 	void useSkillbook(Player *player, ReadPacket *packet);
-	void takeItemSlot(Player *player, char inv, short slot, short amount, bool takeStar = false);
+	void takeItemSlot(Player *player, char inv, int16_t slot, int16_t amount, bool takeStar = false);
 	void useChair(Player *player, ReadPacket *packet);
 	void useItemEffect(Player *player, ReadPacket *packet);
 	void stopChair(Player *player, ReadPacket *packet);
